requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
aiohttp-cors==0.7.0
alabaster==0.7.8
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.1
blinker==1.4
bottle==0.12.18
Brlapi==0.7.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
cloudpickle==1.3.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cssselect2==0.3.0
cupshelpers==1.0
cycler==0.10.0
daphne==2.5.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.8
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-formtools==2.2
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gns3-gui==2.2.11
gns3-server==2.2.11
html5lib==1.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
ikpy==3.0.1
imagesize==1.2.0
importlib-metadata==1.5.0
incremental==17.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.2
Jinja2==2.10.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
keyring==18.0.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
logilab-common==1.4.3
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
mpmath==1.1.0
multidict==4.7.6
mysqlclient==2.0.1
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
notebook==6.1.4
numpy==1.19.1
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.4.0.42
packaging==20.3
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
poppy-creature==2.0.0
poppy-ergo-jr==2.0.0
prometheus-client==0.8.0
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.7.0
ptyprocess==0.6.0
py-cpuinfo==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.20
pycups==1.9.73
pyflakes==2.1.1
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.3
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.9.5
pypot==4.0.3
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
pysvn==1.9.9
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
QtAwesome==0.4.4
qtconsole==4.6.0
QtPy==1.9.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
rope==0.10.5
scipy==1.5.2
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
sentry-sdk==0.17.8
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==1.9.5
Sphinx==1.8.5
spyder==3.3.6
spyder-kernels==0.5.2
sqlparse==0.3.1
sympy==1.6.2
systemd-python==234
terminado==0.9.1
testpath==0.4.4
tinycss2==1.0.2
toml==0.10.1
tornado==5.1.1
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.1.8
WeasyPrint==51
webencodings==0.5.1
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xhtml2pdf==0.2.4
xkit==0.0.0
yarl==1.6.0
zipp==1.0.0
zope.interface==5.1.0
